requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
brine==1.0.0
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
future==0.17.1
graphviz==0.8.4           # via mxnet-cu100
idna==2.8                 # via requests
kiwisolver==1.1.0         # via matplotlib
matplotlib==3.1.0
mxnet-cu100==1.4.1
numpy==1.14.6
pillow==6.0.0             # via torchvision
pyparsing==2.4.0          # via matplotlib
python-dateutil==2.8.0    # via matplotlib
requests==2.22.0          # via mxnet-cu100
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
urllib3==1.25.3           # via requests
